+++ test/json_cache_cross_local_storage_test.dart
@@ -21,6 +21,31 @@ void main() {
       await prefs.clear();
     });
 
+    test('contains', () async {
+      const profKey = 'profile';
+      const prefKey = 'preferences';
+      final profData = <String, dynamic>{'id': 1, 'name': 'John Due'};
+      final prefData = <String, dynamic>{
+        'theme': 'dark',
+        'notifications': {'enabled': true}
+      };
+      final crossLocal =
+          JsonCacheCrossLocalStorage(await LocalStorage.getInstance());
+      // update data
+      await crossLocal.refresh(profKey, profData);
+      await crossLocal.refresh(prefKey, prefData);
+
+      // test for `true`
+      expect(await crossLocal.contains(profKey), true);
+      expect(await crossLocal.contains(prefKey), true);
+
+      // test for `false`
+      expect(await crossLocal.contains('a key'), false);
+      await crossLocal.remove(profKey);
+      expect(await crossLocal.contains(profKey), false);
+      await crossLocal.remove(prefKey);
+      expect(await crossLocal.contains(prefKey), false);
+    });
     test('remove', () async {
       final prefs = await LocalStorage.getInstance();
       const profKey = 'profile';