+++ test/json_cache_mem_test.dart
@@ -110,6 +110,26 @@ void main() {
         expect(copy.isEmpty, true);
       });
     });
+    test('contains', () async {
+      final profData = <String, dynamic>{'id': 1, 'name': 'John Due'};
+      final prefData = <String, dynamic>{
+        'theme': 'dark',
+        'notifications': {'enabled': true}
+      };
+      final memCache = JsonCacheMem();
+      // update data
+      await memCache.refresh(profKey, profData);
+      await memCache.refresh(prefKey, prefData);
+
+      // test for `true`
+      expect(await memCache.contains(profKey), true);
+      expect(await memCache.contains(prefKey), true);
+
+      // test for `false`
+      await memCache.remove(prefKey);
+      expect(await memCache.contains(prefKey), false);
+      expect(await memCache.contains('a key'), false);
+    });
     group('remove', () {
       test('default ctor', () async {
         final JsonCacheMem memCache = JsonCacheMem();